+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - gen_libs.pascalize:  Pascal cases a string.
 
 ### Removed
+- cmds_gen.run_prog function.
 - machine.Solaris class.