+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - gen_libs.pascalize:  Pascal cases a string.
 
 ### Removed
+- cmds_gen.get_inst function
 - cmds_gen.create_cfg_array function
 - cmds_gen.run_prog function
 - machine.Solaris class